requirements.txt
boto3==1.7.42
botocore==1.10.42
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
futures==3.2.0
idna==2.6
jmespath==0.9.3
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
rollbar==0.14.0
s3transfer==0.1.13
six==1.11.0
slackclient==1.2.1
timeago==1.0.8
urllib3==1.22
WebOb==1.8.2
websocket-client==0.48.0
